requirements.txt
langchain==1.2.9
langchain-core==1.2.16
langchain-community==0.3.31
langchain-groq==0.1.0
requests==2.32.5
pandas==2.2.3
python-dateutil==2.9.0
openpyxl==3.1.5
streamlit==1.42.0
python-dotenv==1.0.1
plotly==5.23.0
